Create new page blogs and move Navbar to core
