+++ pages/blogs.js
@@ -1,0 +1,11 @@
+import React, { Fragment } from 'react'
+import Head from '~/views/blogs/components/Head'
+import Navbar from '~/views/core/components/Navbar'
+const Blogs = () => (
+  <Fragment>
+    <Head />
+    <Navbar />
+  </Fragment>
+)
+
+export default Blogs
