+++ views/blogs/components/Head.js
@@ -1,0 +1,13 @@
+import React from 'react'
+import NextHead from 'next/head'
+
+const Head = () => (
+  <NextHead>
+    <title>Coding with the Devil ;)</title>
+    <meta charSet="utf-8" />
+    <meta name="description" content="Coding with the Devil ;)" />
+    <meta name="keywords" content="coding, blogs, diaries" />
+  </NextHead>
+)
+
+export default Head
