+++ views/core/components/nav/NavContent.js
@@ -1,8 +1,8 @@
 import React, { useState, useEffect } from 'react'
 import styled from 'styled-components/macro'
 import PropTypes from 'prop-types'
-import NavLogo from '~/views/home/components/nav/NavLogo'
-import NavItem from '~/views/home/components/nav/NavItem'
+import NavLogo from '~/views/core/components/nav/NavLogo'
+import NavItem from '~/views/core/components/nav/NavItem'
 import MobileContainer from '~/views/home/components/common/container/MobileContainer'
 import Row from 'react-bootstrap/Row'
 import Col from 'react-bootstrap/Col'
