+++ views/core/components/nav/NavHeader.js
@@ -4,7 +4,7 @@ import PropTypes from 'prop-types'
 import { Translate } from 'react-localize-redux'
 import { GiHamburgerMenu } from 'react-icons/gi'
 import { MdClose } from 'react-icons/md'
-import NavLang from '~/views/home/components/nav/NavLang'
+import NavLang from '~/views/core/components/nav/NavLang'
 import MobileContainer from '~/views/home/components/common/container/MobileContainer'
 import Row from 'react-bootstrap/Row'
 import Col from 'react-bootstrap/Col'
