+++ views/core/components/nav/NavLogo.js
