+++ views/core/components/nav/__test__/NavLang.spec.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
 import { shallowWithTheme } from '~/utils/withThemeProviders'
-import NavLang from '~/views/home/components/nav/NavLang'
+import NavLang from '~/views/core/components/nav/NavLang'
 
 describe('Nav Lang', () => {
